Creating new test data files and replacing refs to AdminApiTestUtil - saving progress
