+++ server/src/test/java/org/tctalent/server/api/admin/CandidateCitizenshipAdminApiTest.java
@@ -42,6 +42,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.tctalent.server.data.CandidateTestData;
 import org.tctalent.server.model.db.CandidateCitizenship;
 import org.tctalent.server.request.candidate.citizenship.CreateCandidateCitizenshipRequest;
 import org.tctalent.server.service.db.CandidateCitizenshipService;
@@ -59,7 +60,8 @@ class CandidateCitizenshipAdminApiTest extends ApiTestBase {
 
     private static final long CANDIDATE_ID = 99L;
 
-    private final CandidateCitizenship candidateCitizenship = AdminApiTestUtil.getCandidateCitizenship();
+    private final CandidateCitizenship candidateCitizenship =
+        CandidateTestData.getCandidateCitizenship();
 
 
     @MockBean
