+++ server/src/test/java/org/tctalent/server/api/admin/CandidateDependantAdminApiTest.java
@@ -27,6 +27,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.tctalent.server.data.CandidateTestData;
 import org.tctalent.server.model.db.Candidate;
 import org.tctalent.server.model.db.CandidateDependant;
 import org.tctalent.server.request.candidate.dependant.CreateCandidateDependantRequest;
@@ -59,8 +60,8 @@ class CandidateDependantAdminApiTest extends ApiTestBase {
     private static final String BASE_PATH = "/api/admin/candidate-dependant";
     private static final long CANDIDATE_ID = 99L;
 
-    private final Candidate candidate = AdminApiTestUtil.getCandidate();
-    private final CandidateDependant candidateDependant = AdminApiTestUtil.getCandidateDependant();
+    private final Candidate candidate = CandidateTestData.getCandidate();
+    private final CandidateDependant candidateDependant = CandidateTestData.getCandidateDependant();
 
     @MockBean CandidateDependantService candidateDependantService;
     @MockBean CandidateService candidateService;
