+++ server/src/test/java/org/tctalent/server/api/admin/CandidateExamAdminApiTest.java
@@ -26,6 +26,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.tctalent.server.data.CandidateTestData;
 import org.tctalent.server.model.db.CandidateExam;
 import org.tctalent.server.request.candidate.exam.CreateCandidateExamRequest;
 import org.tctalent.server.service.db.CandidateExamService;
@@ -57,7 +58,7 @@ class CandidateExamAdminApiTest extends ApiTestBase {
     private static final String BASE_PATH = "/api/admin/candidate-exam";
     private static final long CANDIDATE_ID = 99L;
 
-    private final CandidateExam candidateExam = AdminApiTestUtil.getCandidateExam();
+    private final CandidateExam candidateExam = CandidateTestData.getCandidateExam();
 
     @MockBean CandidateExamService candidateExamService;
     @MockBean CandidateService candidateService;
