+++ server/src/test/java/org/tctalent/server/api/admin/CandidateJobExperienceAdminApiTest.java
@@ -48,6 +48,7 @@ import org.springframework.data.domain.PageRequest;
 import org.springframework.data.domain.Sort;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.tctalent.server.data.CandidateTestData;
 import org.tctalent.server.model.db.CandidateJobExperience;
 import org.tctalent.server.request.work.experience.CreateJobExperienceRequest;
 import org.tctalent.server.request.work.experience.SearchJobExperienceRequest;
@@ -69,7 +70,8 @@ class CandidateJobExperienceAdminApiTest extends ApiTestBase {
     private static final String BASE_PATH = "/api/admin/candidate-job-experience";
     private static final String SEARCH_PATH = "/search";
 
-    private static final CandidateJobExperience candidateJobExperience = AdminApiTestUtil.getCandidateJobExperience();
+    private static final CandidateJobExperience candidateJobExperience =
+        CandidateTestData.getCandidateJobExperience();
 
     private final Page<CandidateJobExperience> candidateJobExperiencePage =
             new PageImpl<>(
