+++ server/src/test/java/org/tctalent/server/api/admin/CandidateLanguageAdminApiTest.java
@@ -26,6 +26,7 @@ import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.http.MediaType;
 import org.springframework.test.web.servlet.MockMvc;
+import org.tctalent.server.data.CandidateTestData;
 import org.tctalent.server.model.db.CandidateLanguage;
 import org.tctalent.server.request.candidate.language.CreateCandidateLanguageRequest;
 import org.tctalent.server.request.candidate.language.UpdateCandidateLanguageRequest;
@@ -64,9 +65,11 @@ class CandidateLanguageAdminApiTest extends ApiTestBase {
     private static final String BASE_PATH = "/api/admin/candidate-language";
     private static final String GET_LIST_PATH = "/{id}/list";
 
-    private static final CandidateLanguage candidateLanguage = AdminApiTestUtil.getCandidateLanguage();
+    private static final CandidateLanguage candidateLanguage =
+        CandidateTestData.getCandidateLanguage();
 
-    private static final List<CandidateLanguage> candidateLanguageList = AdminApiTestUtil.getListOfCandidateLanguages();
+    private static final List<CandidateLanguage> candidateLanguageList =
+        CandidateTestData.getListOfCandidateLanguages();
 
     @MockBean CandidateLanguageService candidateLanguageService;
