+++ server/src/test/java/org/tctalent/server/api/admin/CandidateSkillAdminApiTest.java
@@ -48,7 +48,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
-import static org.tctalent.server.api.admin.AdminApiTestUtil.getCandidateSkill;
+import static org.tctalent.server.data.CandidateTestData.getCandidateSkill;
 
 /**
  * Unit tests for Candidate Visa Job Check Admin Api endpoints.
