+++ server/src/test/java/org/tctalent/server/api/admin/EducationLevelAdminApiTest.java
@@ -35,6 +35,7 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.tctalent.server.data.LanguageTestData.getSystemLanguage;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import java.io.InputStream;
@@ -77,7 +78,7 @@ class EducationLevelAdminApiTest extends ApiTestBase {
   private static final String SYSTEM_LANGUAGE_TRANSLATION_PATH = "/system/{langCode}";
   private static final String SEARCH_PAGED_PATH = "/search";
 
-  private final SystemLanguage systemLanguage = AdminApiTestUtil.getSystemLanguage();
+  private final SystemLanguage systemLanguage = getSystemLanguage();
   private final List<EducationLevel> educationLevels = AdminApiTestUtil.getEducationLevels();
 
   private final Page<EducationLevel> educationLevelPage =
