+++ server/src/test/java/org/tctalent/server/api/admin/SalesforceAdminApiTest.java
@@ -29,6 +29,7 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.tctalent.server.data.OpportunityTestData.getOpportunity;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 import org.junit.jupiter.api.BeforeEach;
@@ -61,7 +62,7 @@ class SalesforceAdminApiTest extends ApiTestBase {
   private static final String SF_OPPORTUNITY_PATH = "/opportunity";
   private static final String UPDATE_EMPLOYER_OPPORTUNITY_PATH = "/update-emp-opp";
 
-  private final Opportunity opportunity = AdminApiTestUtil.getSalesforceOpportunity();
+  private final Opportunity opportunity = getOpportunity();
 
   @MockBean SalesforceService salesforceService;
   @MockBean SalesforceRecordTypeConfig salesforceRecordTypeConfig;
@@ -99,7 +100,7 @@ class SalesforceAdminApiTest extends ApiTestBase {
 
         .andDo(print())
         .andExpect(status().isOk())
-        .andExpect(jsonPath("$.name", is("SF Opportunity")));
+        .andExpect(jsonPath("$.name", is("Opportunity")));
 
     verify(salesforceService).fetchJobOpportunity(anyString());
   }
