+++ server/src/test/java/org/tctalent/server/api/admin/SavedSearchAdminApiTest.java
@@ -32,6 +32,8 @@ import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.content;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.jsonPath;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.status;
+import static org.tctalent.server.data.SavedListTestData.getSavedList;
+import static org.tctalent.server.data.SavedSearchTestData.getSavedSearch;
 import static org.tctalent.server.model.db.SavedSearchType.other;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
@@ -90,8 +92,8 @@ class SavedSearchAdminApiTest extends ApiTestBase {
   private static final String BASE_PATH = "/api/admin/saved-search";
   private static final String SEARCH_PATH = "/search";
   private static final String SEARCH_PAGED_PATH = "/search-paged";
-  private static final SavedSearch savedSearch = AdminApiTestUtil.getSavedSearch();
-  private static final SavedList savedList = AdminApiTestUtil.getSavedList();
+  private static final SavedSearch savedSearch = getSavedSearch();
+  private static final SavedList savedList = getSavedList();
   private static final List<SavedSearch> savedSearchList = AdminApiTestUtil.getListOfSavedSearches();
   private static final String CLEAR_SELECTION_PATH = "/clear-selection/";
   private static final String CREATE_FROM_DEFAULT_PATH = "/create-from-default";
