+++ server/src/test/java/org/tctalent/server/api/admin/TranslationAdminApiTest.java
@@ -58,6 +58,10 @@ import static org.springframework.test.web.servlet.request.MockMvcRequestBuilder
 import static org.springframework.test.web.servlet.result.MockMvcResultHandlers.print;
 import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.*;
 import static org.tctalent.server.api.admin.AdminApiTestUtil.getTranslationFile;
+import static org.tctalent.server.data.CandidateTestData.getListOfOccupations;
+import static org.tctalent.server.data.CountryTestData.getSourceCountryListA;
+import static org.tctalent.server.data.LanguageTestData.getLanguageLevelList;
+import static org.tctalent.server.data.LanguageTestData.getLanguageList;
 
 /**
  * Unit tests for Translation Admin Api endpoints.
@@ -80,25 +84,25 @@ class TranslationAdminApiTest extends ApiTestBase {
 
   private static final Translation translation = AdminApiTestUtil.getTranslation();
   private final Page<Country> countryPage = new PageImpl<>(
-          AdminApiTestUtil.getCountries(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      getSourceCountryListA(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   private final Page<Language> languagePage = new PageImpl<>(
-          AdminApiTestUtil.getLanguageList(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      getLanguageList(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   private final Page<LanguageLevel> languageLevelPage = new PageImpl<>(
-          AdminApiTestUtil.getLanguageLevelList(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      getLanguageLevelList(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   private final Page<Occupation> occupationPage = new PageImpl<>(
-          AdminApiTestUtil.getListOfOccupations(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      getListOfOccupations(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   private final Page<EducationLevel> educationLevelPage = new PageImpl<>(
-          AdminApiTestUtil.getEducationLevels(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      AdminApiTestUtil.getEducationLevels(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   private final Page<EducationMajor> educationMajorPage = new PageImpl<>(
-          AdminApiTestUtil.getEducationMajors(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      AdminApiTestUtil.getEducationMajors(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   private final Page<SurveyType> surveyTypePage = new PageImpl<>(
-          AdminApiTestUtil.getSurveyTypes(), PageRequest.of(0, 10, Sort.unsorted()), 1);
+      AdminApiTestUtil.getSurveyTypes(), PageRequest.of(0, 10, Sort.unsorted()), 1);
 
   @MockBean
   TranslationService translationService;
@@ -155,15 +159,15 @@ class TranslationAdminApiTest extends ApiTestBase {
             .andDo(print())
             .andExpect(status().isOk())
             .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.totalElements", is(3)))
+            .andExpect(jsonPath("$.totalElements", is(2)))
             .andExpect(jsonPath("$.totalPages", is(1)))
             .andExpect(jsonPath("$.number", is(0)))
             .andExpect(jsonPath("$.hasNext", is(false)))
             .andExpect(jsonPath("$.hasPrevious", is(false)))
             .andExpect(jsonPath("$.content", notNullValue()))
-            .andExpect(jsonPath("$.content.[0].name", is("Jordan")))
+            .andExpect(jsonPath("$.content.[0].name", is("Lebanon")))
             .andExpect(jsonPath("$.content.[0].status", is("active")))
-            .andExpect(jsonPath("$.content.[0].translatedName", is("Jordan")));
+            .andExpect(jsonPath("$.content.[0].translatedName", is("Lebanon")));
 
     verify(countryService).searchCountries(any(SearchCountryRequest.class));
   }
@@ -187,15 +191,15 @@ class TranslationAdminApiTest extends ApiTestBase {
             .andDo(print())
             .andExpect(status().isOk())
             .andExpect(content().contentType(MediaType.APPLICATION_JSON))
-            .andExpect(jsonPath("$.totalElements", is(3)))
+            .andExpect(jsonPath("$.totalElements", is(2)))
             .andExpect(jsonPath("$.totalPages", is(1)))
             .andExpect(jsonPath("$.number", is(0)))
             .andExpect(jsonPath("$.hasNext", is(false)))
             .andExpect(jsonPath("$.hasPrevious", is(false)))
             .andExpect(jsonPath("$.content", notNullValue()))
-            .andExpect(jsonPath("$.content.[0].name", is("Jordan")))
+            .andExpect(jsonPath("$.content.[0].name", is("Lebanon")))
             .andExpect(jsonPath("$.content.[0].status", is("active")))
-            .andExpect(jsonPath("$.content.[0].translatedName", is("Jordan")));
+            .andExpect(jsonPath("$.content.[0].translatedName", is("Lebanon")));
 
     verify(countryService).searchCountries(any(SearchCountryRequest.class));
   }
