+++ server/src/test/java/org/tctalent/server/api/chat/ChatPostAdminApiTest.java
@@ -44,8 +44,8 @@ import org.springframework.http.MediaType;
 import org.springframework.mock.web.MockMultipartFile;
 import org.springframework.test.web.servlet.MockMvc;
 import org.springframework.web.multipart.MultipartFile;
-import org.tctalent.server.api.admin.AdminApiTestUtil;
 import org.tctalent.server.api.admin.ApiTestBase;
+import org.tctalent.server.data.JobChatTestData;
 import org.tctalent.server.model.db.ChatPost;
 import org.tctalent.server.service.db.impl.ChatPostServiceImpl;
 
@@ -58,7 +58,7 @@ class ChatPostAdminApiTest extends ApiTestBase {
     private static final String BASE_PATH = "/api/admin/chat-post";
     private static final String LIST = "/list";
     private static final String UPLOAD = "/upload";
-    private static final List<ChatPost> postList = AdminApiTestUtil.getListOfPosts();
+    private static final List<ChatPost> postList = JobChatTestData.getListOfPosts();
 
     @Autowired
     MockMvc mockMvc;
